requirements.txt
fastapi==0.104.1
urllib3==2.1.0
uvicorn==0.24.0.post1
PyYAML==6.0.1